refactor(node/process): align implementation and cleanup
